refactor: change api folder structure (#16)
